+++ app/controllers/app_configs_controller.rb
@@ -40,7 +40,7 @@ class AppConfigsController < SignedInApplicationController
         :notification_channel,
         :bitrise_project_id,
         :firebase_android_config,
-        :firebase_ios_config,
+        :firebase_ios_config
       )
   end
 
@@ -73,7 +73,7 @@ class AppConfigsController < SignedInApplicationController
 
   def require_integration_setup
     unless @app.app_setup_instructions[:app_config][:visible]
-      redirect_to app_path(@app), flash: { notice: "Finish the integration setup before configuring the app." }
+      redirect_to app_path(@app), flash: {notice: "Finish the integration setup before configuring the app."}
     end
   end
 end
