+++ config/schema/platform_aware_integration_config.json
@@ -9,15 +9,6 @@
     },
     "display_name": {
       "type": ["string", "null"]
-    },
-    "provider": {
-      "type": ["string", "null"]
-    },
-    "repo_url": {
-      "type": ["string", "null"]
-    },
-    "avatar_url": {
-      "type": ["string", "null"]
     }
   }
 }
